v1.0.3
    - Changed typedef for post function hooks.
    - Made the States enum a normal enum instead of an enum class to reduce the need to cast to its underlying type.
    - Added an "Initialized" private member for modules.
    - Made modules initialize first, then function hooks/instances.

+++ Components/Components/GameState.cpp
@@ -25,8 +25,8 @@ std::string GameStateComponent::GetStateLabel()
 
 bool GameStateComponent::IsInGame() const
 {
-	if (GetStateId() & static_cast<uint32_t>(States::CasualMatch)
-		|| GetStateId() & static_cast<uint32_t>(States::RankedMatch))
+	if (GetStateId() & States::CasualMatch
+		|| GetStateId() & States::RankedMatch)
 	{
 		return true;
 	}
